Fix dash and refactor
