+++ src/test/resources/search-container-cases.json
@@ -1,7 +1,6 @@
 [
-  {"id":"BOR-AKSA","description":"a point in the open, its container is a boundary whose polygon self intersects","lat":31.44874,"lng":35.03820,"expectedContainer":"שטח C"},
-  {"id":"MONKS-CAVE","description":"another point whose only container is that same boundary","lat":31.6464,"lng":35.125,"expectedContainer":"שטח C"},
-  {"id":"JERUSALEM","description":"a point inside a city, the container was never broken","lat":31.7683,"lng":35.2137,"expectedContainer":"ירושלים"},
-  {"id":"MATE-YEHUDA","description":"a regional council whose polygon self intersects as well","lat":31.7130,"lng":35.0170,"expectedContainer":"מועצה אזורית מטה יהודה"}
-
+  {"id":"point-in-self-intersecting-boundary","description":"a point in the open, its container is a boundary whose polygon self intersects","lat":31.44874,"lng":35.03820,"expectedContainer":"שטח C"},
+  {"id":"second-point-in-same-self-intersecting-boundary","description":"another point whose only container is that same boundary","lat":31.6464,"lng":35.125,"expectedContainer":"שטח C"},
+  {"id":"point-inside-intact-city-boundary","description":"a point inside a city, the container was never broken","lat":31.7683,"lng":35.2137,"expectedContainer":"ירושלים"},
+  {"id":"point-in-self-intersecting-regional-council","description":"a regional council whose polygon self intersects as well","lat":31.7130,"lng":35.0170,"expectedContainer":"מועצה אזורית מטה יהודה"}
 ]
