+++ src/test/resources/search-sanity-cases.json
@@ -1,20 +1,20 @@
 [
-  {"id":"HAIFA-HE","searchTerm":"חיפה","uiLanguage":"he","center":[32.79,34.99],"zoom":11,"expectedTarget":[32.8191,34.9983],"radiusMeters":15000,"topN":3},
-  {"id":"HAIFA-EN","searchTerm":"Haifa","uiLanguage":"en","center":[32.79,34.99],"zoom":11,"expectedTarget":[32.8191,34.9983],"radiusMeters":15000,"topN":3},
-  {"id":"JERUSALEM-HE","searchTerm":"ירושלים","uiLanguage":"he","center":[31.77,35.21],"zoom":11,"expectedTarget":[31.7683,35.2137],"radiusMeters":15000,"topN":3},
-  {"id":"JERUSALEM-EN","searchTerm":"Jerusalem","uiLanguage":"en","center":[31.77,35.21],"zoom":11,"expectedTarget":[31.7683,35.2137],"radiusMeters":15000,"topN":3},
-  {"id":"TELAVIV-HE","searchTerm":"תל אביב","uiLanguage":"he","center":[32.08,34.78],"zoom":11,"expectedTarget":[32.0879,34.7893],"radiusMeters":15000,"topN":3},
-  {"id":"BEERSHEVA-HE","searchTerm":"באר שבע","uiLanguage":"he","center":[31.25,34.79],"zoom":11,"expectedTarget":[31.2530,34.7915],"radiusMeters":15000,"topN":3},
-  {"id":"EILAT-HE","searchTerm":"אילת","uiLanguage":"he","center":[29.55,34.95],"zoom":11,"expectedTarget":[29.5581,34.9482],"radiusMeters":15000,"topN":3},
-  {"id":"NAZARETH-HE","searchTerm":"נצרת","uiLanguage":"he","center":[32.70,35.30],"zoom":11,"expectedTarget":[32.7028,35.2973],"radiusMeters":15000,"topN":3},
-  {"id":"MASADA-HE","searchTerm":"מצדה","uiLanguage":"he","center":[31.31,35.35],"zoom":12,"expectedTarget":[31.3156,35.3536],"radiusMeters":5000,"topN":3},
-  {"id":"MERON-HE","searchTerm":"הר מירון","uiLanguage":"he","center":[33.00,35.40],"zoom":12,"expectedTarget":[32.9973,35.4046],"radiusMeters":5000,"topN":3},
-  {"id":"NO-CENTER-HE","searchTerm":"ירושלים","uiLanguage":"he","expectedTarget":[31.7683,35.2137],"radiusMeters":15000,"topN":3},
-  {"id":"PREFIX-HE","searchTerm":"ירושל","uiLanguage":"he","center":[31.77,35.21],"zoom":11,"prefix":true,"expectedTarget":[31.7683,35.2137],"radiusMeters":15000,"topN":3},
-  {"id":"EXACT-HE","searchTerm":"חיפה","uiLanguage":"he","template":"points_search_exact","expectedTarget":[32.8191,34.9983],"radiusMeters":15000,"topN":3},
-  {"id":"EXTERNAL","searchTerm":"some name","uiLanguage":"en","expectedTarget":[33.250042,35.729807],"radiusMeters":1000,"topN":3},
-  {"id":"FAR-FROM-CENTER","searchTerm":"טחנת הנזירים","uiLanguage":"he","center":[31.5292,35.0327],"zoom":13,"expectedTarget":[32.7648,35.1569],"radiusMeters":1000,"topN":5},
-  {"id":"FAR-FROM-CENTER-TYPO","searchTerm":"תחנת הנזירים","uiLanguage":"he","center":[31.5292,35.0327],"zoom":13,"expectedTarget":[32.7648,35.1569],"radiusMeters":1000,"topN":5},
-  {"id":"FAR-CITY","searchTerm":"חיפה","uiLanguage":"he","center":[29.55,34.95],"zoom":12,"expectedTarget":[32.8191,34.9983],"radiusMeters":15000,"topN":3},
-  {"id":"FAR-SITE","searchTerm":"מצדה","uiLanguage":"he","center":[32.08,34.78],"zoom":12,"expectedTarget":[31.3156,35.3536],"radiusMeters":5000,"topN":3}
+  {"id":"find-haifa-hebrew","searchTerm":"חיפה","uiLanguage":"he","center":[32.79,34.99],"zoom":11,"expectedTarget":[32.8191,34.9983],"radiusMeters":15000,"topN":3},
+  {"id":"find-haifa-english","searchTerm":"Haifa","uiLanguage":"en","center":[32.79,34.99],"zoom":11,"expectedTarget":[32.8191,34.9983],"radiusMeters":15000,"topN":3},
+  {"id":"find-jerusalem-hebrew","searchTerm":"ירושלים","uiLanguage":"he","center":[31.77,35.21],"zoom":11,"expectedTarget":[31.7683,35.2137],"radiusMeters":15000,"topN":3},
+  {"id":"find-jerusalem-english","searchTerm":"Jerusalem","uiLanguage":"en","center":[31.77,35.21],"zoom":11,"expectedTarget":[31.7683,35.2137],"radiusMeters":15000,"topN":3},
+  {"id":"find-tel-aviv-hebrew","searchTerm":"תל אביב","uiLanguage":"he","center":[32.08,34.78],"zoom":11,"expectedTarget":[32.0879,34.7893],"radiusMeters":15000,"topN":3},
+  {"id":"find-beer-sheva-hebrew","searchTerm":"באר שבע","uiLanguage":"he","center":[31.25,34.79],"zoom":11,"expectedTarget":[31.2530,34.7915],"radiusMeters":15000,"topN":3},
+  {"id":"find-eilat-hebrew","searchTerm":"אילת","uiLanguage":"he","center":[29.55,34.95],"zoom":11,"expectedTarget":[29.5581,34.9482],"radiusMeters":15000,"topN":3},
+  {"id":"find-nazareth-hebrew","searchTerm":"נצרת","uiLanguage":"he","center":[32.70,35.30],"zoom":11,"expectedTarget":[32.7028,35.2973],"radiusMeters":15000,"topN":3},
+  {"id":"find-masada-hebrew","searchTerm":"מצדה","uiLanguage":"he","center":[31.31,35.35],"zoom":12,"expectedTarget":[31.3156,35.3536],"radiusMeters":5000,"topN":3},
+  {"id":"find-mount-meron-hebrew","searchTerm":"הר מירון","uiLanguage":"he","center":[33.00,35.40],"zoom":12,"expectedTarget":[32.9973,35.4046],"radiusMeters":5000,"topN":3},
+  {"id":"find-jerusalem-without-center","searchTerm":"ירושלים","uiLanguage":"he","expectedTarget":[31.7683,35.2137],"radiusMeters":15000,"topN":3},
+  {"id":"find-jerusalem-by-prefix","searchTerm":"ירושל","uiLanguage":"he","center":[31.77,35.21],"zoom":11,"prefix":true,"expectedTarget":[31.7683,35.2137],"radiusMeters":15000,"topN":3},
+  {"id":"find-haifa-with-exact-template","searchTerm":"חיפה","uiLanguage":"he","template":"points_search_exact","expectedTarget":[32.8191,34.9983],"radiusMeters":15000,"topN":3},
+  {"id":"find-external-geojson-point","searchTerm":"some name","uiLanguage":"en","expectedTarget":[33.250042,35.729807],"radiusMeters":1000,"topN":3},
+  {"id":"find-monks-mill-far-from-center","searchTerm":"טחנת הנזירים","uiLanguage":"he","center":[31.5292,35.0327],"zoom":13,"expectedTarget":[32.7648,35.1569],"radiusMeters":1000,"topN":5},
+  {"id":"find-monks-mill-far-from-center-with-typo","searchTerm":"תחנת הנזירים","uiLanguage":"he","center":[31.5292,35.0327],"zoom":13,"expectedTarget":[32.7648,35.1569],"radiusMeters":1000,"topN":5},
+  {"id":"find-haifa-from-distant-center","searchTerm":"חיפה","uiLanguage":"he","center":[29.55,34.95],"zoom":12,"expectedTarget":[32.8191,34.9983],"radiusMeters":15000,"topN":3},
+  {"id":"find-masada-from-distant-center","searchTerm":"מצדה","uiLanguage":"he","center":[32.08,34.78],"zoom":12,"expectedTarget":[31.3156,35.3536],"radiusMeters":5000,"topN":3}
 ]
